\name{countRepeats-package}
\alias{countRepeats-package}
\alias{countRepeats}
\docType{package}
\title{
\packageTitle{countRepeats}
}
\description{
\packageDescription{countRepeats}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{countRepeats}
}
\author{
\packageAuthor{countRepeats}

Maintainer: \packageMaintainer{countRepeats}
}

\keyword{ package }
\seealso{
\code{\link[summarizeOverlaps.adjNH]{countRepeats}} 
}

